requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==2.2.7
django-betterforms==1.2
django-extensions==2.2.9
future==0.18.2
idna==2.9
iso8601==0.1.12
jsonfield==3.1.0
kiwisolver==1.1.0
Pillow==7.0.0
pydot==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
serial==0.0.97
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
tensorflow==2.0.0